move: 테스트파일 디렉토리 정리
Ref: #73
